% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth_object.R
\name{median}
\alias{median}
\title{Median}
\usage{
median(x, ...)
}
\arguments{
\item{x}{An object. For \code{depth} objects, see \code{\link{median.depth}}.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Generic function for computing the median. For \code{depth} objects,
returns the deepest observation. For all other objects, delegates to
\code{stats::median}.
}
